mlops-demo/2-cloud-run-deployment/requirements.txt
torch>=2.0.0
transformers>=4.30.0
flask>=2.0.0
gunicorn>=20.1.0
google-cloud-logging>=3.5.0
google-cloud-monitoring>=2.14.1
python-dotenv>=1.0.0
requests>=2.31.0 